Use Qt magic to parse block geometries.
Rather than having a nasty switch statement which has to be kept in
sync, we now stick the BlockGeometry enum into a QObject and use
QMetaObject to convert between QString and enum constant automatically.
It's generalizable to other enums as well, but the enum in question
needs to be moved to a QObject subclass.  (The object need not actually
be instantiable, in fact it probably shouldn't be.)

Also adds the stub of a TorchRenderable, which should hint at my next
goal...