Save and load block orientations.
Block orientation is now saved and loaded from the .mcdiagram file
format.  The serialization and deserialization is now delegated to
BlockInstance rather than being handled by the diagram.  The file format
version number has been revved, as this change is not
backwards-compatible.

Also makes it possible to construct an invalid BlockInstance, and shares
a header file across between MCModeler and the JSON writer to prevent
the possibility of typos in block data key names.