Prepare support for vertical levels.
This is mostly working already, but face culling is messed up and I'm
not sure why, so it's turned off completely for vertical slices (which
makes it pretty slow).  There's also no place in the save file for this
information to be stored yet.  Accordingly, there's no way to enable
vertical slices without code changes right now.